+++ src/app/forms/radio-button/terra-radio-button.component.ts
@@ -1,8 +1,8 @@
 import {
     Component,
-    Input,
+    forwardRef,
     HostListener,
-    forwardRef
+    Input
 } from '@angular/core';
 import {
     ControlValueAccessor,