+++ src/app/navigator/button-group/data/terra-button-group.interface.ts
@@ -1,0 +1,10 @@
+import { TerraButtonInterface } from '../../../button/data/terra-button.interface';
+
+/**
+ * @author mscharf
+ */
+export interface TerraButtonGroupInterface extends TerraButtonInterface
+{
+    hasChildren:boolean;
+    isVisible:boolean;
+}