+++ src/app/navigator/data/terra-navigator-node.interface.ts
@@ -1,0 +1,13 @@
+/**
+ * @author mscharf
+ */
+export interface TerraNavigatorNodeInterface<D>
+{
+    nodeName:string;
+    nodeIcon?:string;
+    route:string;
+    value?:D;
+    rootPath?:Array<number>;
+    children:Array<TerraNavigatorNodeInterface<D>>;
+    isVisible?:boolean;
+}