+++ src/app/table/data-table/terra-data-table.component.ts
@@ -1,11 +1,11 @@
 import {
     Component,
-    Input,
-    Output,
-    ViewChild,
     EventEmitter,
+    Input,
     OnChanges,
-    SimpleChanges
+    Output,
+    SimpleChanges,
+    ViewChild
 } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { TerraDataTableHeaderCellInterface } from './cell/terra-data-table-header-cell.interface';
@@ -74,7 +74,8 @@ export class TerraDataTableComponent<S extends TerraBaseService, D extends Terra
     {
         if(changes['_hasCheckboxes'])
         {
-            console.warn('_hasCheckboxes is deprecated. It will be removed in one of the upcoming releases. Please use inputHasCheckboxes instead.');
+            console.warn(
+                '_hasCheckboxes is deprecated. It will be removed in one of the upcoming releases. Please use inputHasCheckboxes instead.');
             this.inputHasCheckboxes = changes['_hasCheckboxes'].currentValue;
         }
     }
@@ -107,6 +108,16 @@ export class TerraDataTableComponent<S extends TerraBaseService, D extends Terra
         {
             this.viewChildHeaderCheckbox.isIndeterminate = true;
         }
+    }
+    
+    private checkTooltipPlacement(placement:string):string
+    {
+        if(placement != null && placement != '')
+        {
+            return placement;
+        }
+        
+        return 'top';
     }
     
     private changeRowState(isChecked:boolean,
@@ -142,16 +153,15 @@ export class TerraDataTableComponent<S extends TerraBaseService, D extends Terra
     
     private rowClicked(cell:TerraDataTableCellInterface, row:TerraDataTableRowInterface<D>):void
     {
-        this._rowList
-            .forEach((row) =>
-                     {
-                         row.isActive = false;
-                     });
-        
-        row.isActive = true;
-        
         if(!cell.buttonList)
         {
+            this._rowList
+                .forEach((row) =>
+                         {
+                             row.isActive = false;
+                         });
+            
+            row.isActive = true;
             row.clickFunction();
         }
     }