+++ src/app/tag/data/terra-tag.interface.ts
@@ -4,7 +4,14 @@
  */
 export interface TerraTagInterface
 {
-    badge:string;
     isTagged?:boolean
     isTaggable?:boolean;
+    /**
+     * @deprecated will be removed in next major release. Please use badge instead.
+     */
+    caption?:string;
+    /**
+     *  will be mandatory in next major release.
+     */
+    badge?:string;
 }