+++ src/app/terra-components.module.ts
@@ -1,8 +1,8 @@
 import { TerraLoadingBarService } from './loading-bar/service/terra-loading-bar.service';
 import { BrowserModule } from '@angular/platform-browser';
 import {
-    NgModule,
-    ModuleWithProviders
+    ModuleWithProviders,
+    NgModule
 } from '@angular/core';
 import {
     FormsModule,
@@ -10,11 +10,11 @@ import {
 } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 import {
-    TooltipModule,
     AlertModule,
+    ButtonsModule,
     ModalModule,
-    ButtonsModule
-} from 'ng2-bootstrap';
+    TooltipModule
+} from 'ngx-bootstrap';
 import { TerraComponentsComponent } from './terra-components.component';
 import { TerraTextInputComponent } from './forms/input/text-input/terra-text-input.component';
 import { TerraNumberInputComponent } from './forms/input/number-input/terra-number-input.component';
@@ -56,9 +56,14 @@ import { MyDatePickerModule } from 'mydatepicker';
 import { TerraTextAreaInputComponent } from './forms/input/text-area-input/terra-text-area-input.component';
 import { TerraLoadingSpinnerComponent } from './loading-spinner/terra-loading-spinner.component';
 import { TerraLoadingSpinnerService } from './loading-spinner/service/terra-loading-spinner.service';
+import { TerraCardComponent } from './card/terra-card.component';
 import { TranslationModule } from 'angular-l10n';
 import { TerraUrlParamsDecorator } from './service/data/terra-url-params-decorator.service';
+import { TerraNavigatorComponent } from './navigator/terra-navigator.component';
+import { TerraNavigatorSplitViewConfig } from './navigator/config/terra-navigator-split-view.config';
 import { TerraToggleComponent } from './toggle/terra-toggle.component';
+import { TerraSyntaxEditorComponent } from './editor/syntax/terra-syntax-editor.component';
+import { AceEditorModule } from 'ng2-ace-editor';
 export { TerraAlertPanelComponent } from './alert/terra-alert-panel.component';
 export { TerraAlertComponent } from './alert/terra-alert.component';
 export { TerraButtonInterface } from './button/data/terra-button.interface';
@@ -119,6 +124,9 @@ export { TerraTileBoxColor } from './tile/box/data/terra-tile-box-color';
 export { TerraSuggestionBoxComponent } from './forms/suggestion-box/terra-suggestion-box.component';
 export { TerraDatePickerComponent } from './forms/input/date-picker/terra-date-picker.component';
 export { TerraTextAreaInputComponent } from './forms/input/text-area-input/terra-text-area-input.component';
+export { TerraCardComponent } from './card/terra-card.component';
+export { TerraSyntaxEditorComponent } from './editor/syntax/terra-syntax-editor.component';
+export { TerraSyntaxEditorData } from './editor/syntax/data/terra-syntax-editor.data';
 
 @NgModule({
               declarations:    [
@@ -158,7 +166,10 @@ export { TerraTextAreaInputComponent } from './forms/input/text-area-input/terra
                   TerraSuggestionBoxComponent,
                   TerraDatePickerComponent,
                   TerraTextAreaInputComponent,
-                  TerraToggleComponent
+                  TerraCardComponent,
+                  TerraNavigatorComponent,
+                  TerraToggleComponent,
+                  TerraSyntaxEditorComponent
               ],
               entryComponents: [
                   TerraTextInputComponent,
@@ -192,7 +203,10 @@ export { TerraTextAreaInputComponent } from './forms/input/text-area-input/terra
                   TerraSuggestionBoxComponent,
                   TerraDatePickerComponent,
                   TerraTextAreaInputComponent,
-                  TerraToggleComponent
+                  TerraCardComponent,
+                  TerraNavigatorComponent,
+                  TerraToggleComponent,
+                  TerraSyntaxEditorComponent
               ],
               exports:         [
                   TerraAlertPanelComponent,
@@ -229,7 +243,10 @@ export { TerraTextAreaInputComponent } from './forms/input/text-area-input/terra
                   TerraSuggestionBoxComponent,
                   TerraDatePickerComponent,
                   TerraTextAreaInputComponent,
-                  TerraToggleComponent
+                  TerraCardComponent,
+                  TerraNavigatorComponent,
+                  TerraToggleComponent,
+                  TerraSyntaxEditorComponent
               ],
               imports:         [
                   BrowserModule,
@@ -241,10 +258,12 @@ export { TerraTextAreaInputComponent } from './forms/input/text-area-input/terra
                   AlertModule.forRoot(),
                   ButtonsModule.forRoot(),
                   TranslationModule.forRoot(),
-                  MyDatePickerModule
+                  MyDatePickerModule,
+                  AceEditorModule
               ],
               providers:       [
                   COMPILER_PROVIDERS,
+                  TerraNavigatorSplitViewConfig
               ],
               bootstrap:       [
                   //TerraComponentsComponent
@@ -260,6 +279,7 @@ export class TerraComponentsModule
                         TerraLoadingSpinnerService,
                         TerraDataTableContextMenuService,
                         TerraBaseService,
+                        TerraNavigatorSplitViewConfig,
                         TerraUrlParamsDecorator,
                         TerraAlertComponent]
         };
@@ -273,6 +293,7 @@ export class TerraComponentsModule
                         TerraLoadingSpinnerService,
                         TerraDataTableContextMenuService,
                         TerraBaseService,
+                        TerraNavigatorSplitViewConfig,
                         TerraUrlParamsDecorator,
                         TerraAlertComponent]
         };